DEV-42 Distributes code from Apply function to three components by form step
